pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yt-dlp-getpot-wpc"
version = "0.1.1"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE"}
keywords = ["yt-dlp", "yt-dlp-plugin", "yt-dlp-plugins-get-pot"]
authors = [
  { name = "coletdjnz", email = "coletdjnz@protonmail.com" },
]
dependencies = [ "yt-dlp-get-pot", "nodriver"]

[tool.hatch.env.default]
installer = "uv"
path = ".venv"

[tool.hatch.build.targets.wheel]
packages = ["yt_dlp_plugins"]